Fixed errors numpy vectorisations and also fixed the rotational inference
Firstly numpy vectorisations don't work on the dimensions of a 2D array, but only on the final cells themselves, so we just have to use for loops for now. Since we don't really use much other functions, we can rely on zipping and normal python list operations.

Secondly the clockwise was actually anticlockwise and anticlockwise was actually clockwise. It turns out that the acceleration vector delta directions are the opposite of the rotational tangential direction.